Use APP_KEY instead of MASTER_KEY
when runing the tests using `nosetests` in python 3.x, you'll run
into the problem that it will still look for the APP KEY instead
of using MASTER_KEY, but when you runing the single test, it was
file. And since other test file also using APP_KEY and change to
use APP_KEY do fix the problem, I guess we can use this one.