fix(LeanObject): test existence more accurately
Previously, the code checks whether the object has an truthy objectId,
to determine the existence of a LeanCloud object.
But if the object is constructed via `query.get("nonexist_objectId")`,
then it does have an objectId (`{"objectId": "nonexist_objectId"}`).
Therefore, we check whether the object has the `createdAt` attribute
instead.

Also update the comments of `is_new` and `is_existed`,
to clarify their intented usage.